vaev-style: Replaced f16 by f32 for opacity, because not every platform support f16.

+++ src/vaev-engine/style/specified.cpp
@@ -61,7 +61,7 @@ export struct SpecifiedValues {
     Integer order;
     AlignProps aligns;
     Display display;
-    f16 opacity;
+    f32 opacity;
 
     // Small Field
     Float float_ = Float::NONE;
